ai_models/requirements.txt
# HuggingFace & Transformers
transformers>=4.36.0
huggingface_hub>=0.20.0
datasets>=2.16.0
tokenizers>=0.15.0
accelerate>=0.25.0

# PyTorch (CPU - replace with CUDA version if you have a GPU)
torch>=2.1.0
torchvision>=0.16.0

# Flask API Server
flask>=3.0.0
flask-cors>=4.0.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
colorama>=0.4.6
